// client/src/types.ts
export interface User {
  username: string;
  uid?: string;
}
// ...
export interface UserSettings {
  globalMode: string | null;
  headerMode: string | null;
  sidebarMode: string | null;
  textMode: string | null;
  buttonMode: string | null;
}